Unify test suite for tokenizer
